feature(rewrite): Use the new RxJS Library (#294)
<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->
Utilize the new `ultimate-govee` RxJS library

## Related Issues
Fixes #94 

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
This project was one of the first NestJS projects I had built and the
first Homebridge plugin, as it grew I found a lot of pain points and
issues that needed to be addressed. This PR is a complete rewrite of the
plugin and the core Govee logic has been moved to it's own repository
[constructorfleet/ultimate-govee](https://github.com/constructorfleet/ultimate-govee).

There is a configuration migration process that should migrate existing
configurations to the new schema, but it may not be perfect.

This change properly supports AWS IoT Core and BLE communication, I aim
to add LAN and the new OpenAPI development platform soon.

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->
- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.